+++ GDX/Tests/Editor/StringExtensionsTests.cs
@@ -1,4 +1,7 @@
-﻿using NUnit.Framework;
+﻿// dotBunny licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using NUnit.Framework;
 #if GDX_TESTFRAMEWORK_PERFORMANCETESTING
 using Unity.PerformanceTesting;
 
@@ -9,6 +12,9 @@ using Unity.PerformanceTesting;
 
 namespace GDX.Tests.Editor
 {
+    /// <summary>
+    /// A collection of unit tests to validate functionality of the <see cref="StringExtensions"/> class.
+    /// </summary>
     public class StringExtensionsTests
     {
         private const int WarmupCount = 10;